requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.1
python-dotenv==1.1.1
tomli==2.2.1
typing_extensions==4.14.0
Werkzeug==3.1.3
